move tests down into feature subdir to match production package.
